feat[PanamaUring] 1,异步IOSQE_BUFFER_SELECT read实现 2,BufRIng实现

+++ native-lib/src/main/java/top/dreamlike/helper/StackValue.java
@@ -55,7 +55,7 @@ public class StackValue implements SegmentAllocator, AutoCloseable {
         var base = stack.address();
         var start = alignUp(base + hasAllocate, byteAlignment) - base;
         MemorySegment res = stack.asSlice(start, byteSize, byteAlignment);
-        hasAllocate += start + byteSize;
+        hasAllocate += (int) (start + byteSize);
         return res;
     }
